declare qna types in qna module
